Add -version flag
The new -version flag prints useful information for identifying the
binary version. A nice side-effect of this patch is that it also fixes
`squid_exporter_build_info` to have the branch, revision and version,
which were empty before.

a `VERSION` file was introduced to mark the version in filesystem, as
you can't get the version in compile-time from git without actually
tagging it. It's a kind of chicken-and-egg problem. This solution was
taken from how other projects in the Prometheus ecosystem manage this
problem.