Fix for: Repetitive errors on Time Percentiles #44
Problem: As per the bug report: each time the exporter counters a
"header" type of line while parsing service times it outputs a warning.

In particular the line:

"Service Time Percentiles 5 min 60 min:" causes the warning.

Solution: Make the parsing function decodeServiceTimeStrings ignore
"header" type lines, which are defined as those that end with a ':' at
the end of the line.